Feature idea: Allow custom "login required" handling
My scenario is that I'm using Auth0, so I'm trying to avoid showing a /login screen that just then sends them off to Auth0, it's too jarring.

What I'm doing on my nav login button is simply `$auth.login()`, and I think if a user hits a page requiring auth and aren't, it'd be nice to be able to configure the middleware to do the same, like so:
```javascript
// nuxt.config.js
auth: {
  redirect: {
    login: (ctx) => { ctx.app.$auth.login() }
  }
}
```

I can think of other scenarios where this could be useful, e.g. if a developer needed to put some things in session, send them to login and then when they come back logged in be able to finish processing whatever needed to be done with that session data.