remove default auth0 audience
As of June 8th, the jwt-bearer grant isn't available to new applications. Therefore, any new app cannot get a token without a specified audience (#176). This is a breaking change from upstream.

The audience *must* match the API's audience. However, audience can be omited if a default audience is specified in the tenent's settings. https://auth0.com/docs/tokens/id-token#validate-the-claims