Add error listeners before Auth.init()
In the current setup, if you want to create an error listener, you
should create a plugin, and load them through the config with the
setting auth.plugins, like so:

`plugins: ['~/plugins/auth-error-listener.js']`

The plugins are loaded *after* the main Auth plugin itself is loaded.
The main plugin takes care of the init. So if an error occurs during
init, it's impossible to catch it with a custom error listener.

This commit adds a new config option, called `errorListeners`, it's used
in the same way as plugins, so:

`errorListeners: ['~/utils/auth-error-listener.js']`

The plugin itself isn't a nuxt plugin, so it should export a
default function:

`
export default function(error) {
  console.log('error listener')
  console.log(error.response.data)
  console.log(error.response.status)
}
`

ErrorListeners will be loaded *before* init. This change should be
backwards compatible, since the old method of loading listeners through
plugins still works.