Add video tutorial!
Hey _pi0_!  

I made the video tutorial we talked about.  Hopefully this current version will work, but if there are changes you'd like to see then let me know - I can re-cut the video as needed to make it work better for embedding.

Thanks so much for all the work you do on this module!