Fix #275
A clean, minimal but somewhat incomplete fix for #275.

As per this, we need to specify the base path ( ie, `router.base` ) in the configuration section of auth middleware.
Ie, the configuration `auth.redirect.base` should hold the same value as that of `router.base`

I don't know how to automate this. But this solution will work if we add that config.

Please note that, this fix will not break any existing functionality because, we are using safe default values for this new config option


Signed-off-by: Harish Karumuthil <harish2704@gmail.com>