fix(refresh scheme): fix non-required refresh token flow
Set refresh token to `true` instead of `false` if refresh token is not required. That way refresh token expiration is set. And add `isRefreshing` and `updateOnRefresh` params to be able to not update refresh token expiration if refresh endpoint doesn't refresh the refresh token.
Also simplify `check` as refresh token will always be `true` when not required.