Fix for token.required option not taken into account in refresh scheme
Hi,

When using the refresh scheme with httponly cookies for storing the access and refresh tokens (so the tokens are not even available in the json response in this case), login fails because token.required option is not taken into account, so the scheme still expects a token. This change solves the problem for me when using httponly cookies, I hope it helps.